requirements.txt
Django==5.2.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
django-filter==23.5
social-auth-app-django==5.4.0
Pillow
drf-yasg==1.21.7
django-model-utils==4.3.1
python-dotenv==1.0.0

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
whitenoise==6.6.0
django-cors-headers==4.3.1

# Development dependencies (optional)
django-debug-toolbar==4.2.0